requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.38.0
accelerate>=0.28.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
pycocotools>=2.0.7
scikit-learn>=1.3.0
matplotlib>=3.7.0
requests>=2.31.0
roboflow>=1.1.60
